remove .SILENT from Makefiles, use "make -s" instead
Remove .SILENT from Makefiles, so that full commands are
printed as they are executed. Add "@" in front of "echo" commands
so that echo messages are not printed twice. If it is desired
to suppress printing of the commands, use the "-s" (--quiet)
option of make, for example "make -s all".